financial-valuation-app/backend/pyproject.toml
[tool.poetry]
name = "financial-valuation-backend"
version = "0.1.0"
description = "Flask backend for financial valuation application"
authors = ["Your Name <your.email@example.com>"]
packages = []
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
flask = "^2.3.3"
flask-cors = "^4.0.0"
flask-swagger-ui = "^5.21.0"
apispec = "^6.8.2"
# Pin marshmallow to v3 for compatibility with flask-marshmallow
marshmallow = "^3.20.1"
flask-sqlalchemy = "^3.0.5"
flask-migrate = "^4.0.5"
flask-marshmallow = "^0.15.0"
marshmallow-sqlalchemy = "^0.29.0"
# psycopg2-binary = "^2.9.7"  # Commented out - using SQLite for development
python-dotenv = "^1.0.0"
gunicorn = "^21.2.0"
numpy = "^1.24.3"
pandas = "^2.0.3"
celery = "^5.3.1"
redis = "^4.6.0"
requests = "^2.31.0"
flower = "^2.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
pytest-flask = "^1.2.0"
black = "^23.7.0"
flake8 = "^6.0.0"
mypy = "^1.5.1"
pre-commit = "^3.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true 